working on the urls for admins to delete another user
