--- src/Components/user/nonadmin.tsx
@@ -1,1 +1,0 @@
-export {}
